+++ pkg/webui/containers/collaborator-form/index.js
@@ -14,16 +14,18 @@
 
 import React, { useCallback, useState } from 'react'
 import { useNavigate } from 'react-router-dom'
+import { defineMessages } from 'react-intl'
 
 import Form from '@ttn-lw/components/form'
-import Input from '@ttn-lw/components/input'
 import Notification from '@ttn-lw/components/notification'
-import Radio from '@ttn-lw/components/radio-button'
 import SubmitBar from '@ttn-lw/components/submit-bar'
 import SubmitButton from '@ttn-lw/components/submit-button'
 import toast from '@ttn-lw/components/toast'
 import ModalButton from '@ttn-lw/components/button/modal-button'
 import RightsGroup from '@ttn-lw/components/rights-group'
+
+import CollaboratorSelect from '@console/containers/collaborator-select'
+import composeOption from '@console/containers/collaborator-select/util'
 
 import Yup from '@ttn-lw/lib/yup'
 import PropTypes from '@ttn-lw/lib/prop-types'
@@ -32,13 +34,45 @@ import { userId as collaboratorIdRegexp } from '@ttn-lw/lib/regexp'
 
 import useCollaboratorData from './hooks'
 
+const collaboratorOrganizationSchema = Yup.object().shape({
+  organization_id: Yup.string().matches(collaboratorIdRegexp, sharedMessages.validateAlphanum),
+})
+
+const collaboratorUserSchema = Yup.object().shape({
+  user_id: Yup.string().matches(collaboratorIdRegexp, sharedMessages.validateAlphanum),
+})
+
 const validationSchema = Yup.object().shape({
-  collaborator_id: Yup.string()
-    .matches(collaboratorIdRegexp, Yup.passValues(sharedMessages.validateIdFormat))
-    .required(sharedMessages.validateRequired),
-  collaborator_type: Yup.string().required(sharedMessages.validateRequired),
+  collaborator: Yup.object()
+    .shape({
+      ids: Yup.object().when(['organization_ids'], {
+        is: organizationIds => Boolean(organizationIds),
+        then: collaboratorOrganizationSchema,
+        otherwise: collaboratorUserSchema,
+      }),
+    })
+    .required(sharedMessages.validateRequired)
+    .nullable(),
   rights: Yup.array().min(1, sharedMessages.validateRights),
 })
+
+const m = defineMessages({
+  collaboratorIdPlaceholder: 'Type to choose a collaborator',
+})
+
+const encodeCollaborator = collaboratorOption =>
+  collaboratorOption
+    ? {
+        ids: {
+          [`${collaboratorOption.icon}_ids`]: {
+            [`${collaboratorOption.icon}_id`]: collaboratorOption.value,
+          },
+        },
+      }
+    : null
+
+const decodeCollaborator = collaborator =>
+  collaborator && collaborator.ids ? composeOption(collaborator) : null
 
 const CollaboratorForm = props => {
   const { entity, entityId, collaboratorId, deleteDisabled, update, tts } = props
@@ -63,23 +97,17 @@ const CollaboratorForm = props => {
 
   const handleSubmit = useCallback(
     async (values, { resetForm, setSubmitting }) => {
-      const { collaborator_id, collaborator_type, rights } = values
+      const { collaborator, rights } = values
 
-      const collaborator_ids = {
-        [`${collaborator_type}_ids`]: {
-          [`${collaborator_type}_id`]: collaborator_id,
-        },
-      }
-
-      const collaborator = {
-        ids: collaborator_ids,
+      const composedCollaborator = {
+        ...collaborator,
         rights,
       }
 
       setSubmitError(undefined)
 
       try {
-        await updateCollaborator(collaborator)
+        await updateCollaborator(composedCollaborator)
 
         resetForm({ values })
         if (!update) {
@@ -157,25 +185,18 @@ const CollaboratorForm = props => {
       validationSchema={validationSchema}
     >
       {warning}
-      <Form.Field
-        name="collaborator_id"
-        component={Input}
-        title={sharedMessages.collaboratorId}
-        placeholder={sharedMessages.collaboratorIdPlaceholder}
+      <CollaboratorSelect
+        name="collaborator"
+        title={sharedMessages.collaborator}
+        placeholder={m.collaboratorIdPlaceholder}
         required
         autoFocus={!update}
         disabled={update}
+        entity={entity}
+        entityId={entityId}
+        encode={encodeCollaborator}
+        decode={decodeCollaborator}
       />
-      <Form.Field
-        name="collaborator_type"
-        title={sharedMessages.type}
-        component={Radio.Group}
-        disabled={update}
-        required
-      >
-        <Radio label={sharedMessages.user} value="user" />
-        <Radio label={sharedMessages.organization} value="organization" />
-      </Form.Field>
       <Form.Field
         name="rights"
         title={sharedMessages.rights}